No tengo manera de probar si anda o no modificaciones sobre el detalle. Base para la consulta de inventario

+++ app/src/main/java/domain/Laberinto.java
@@ -9,12 +9,12 @@ public class Laberinto implements Serializable {
 
     String nombre;
     String descripcion;
-    Long id;
+    Integer id;
 
     public Laberinto(){
     }
 
-    public Long getId() {return id;}
+    public Integer getId() {return id;}
 
     public String getNombre() {
         return this.nombre;
